requirements-dev.txt
opencv-python
numpy
Pillow
progress
tqdm